api/requirements.txt
Flask
mysql-connector-python
google-cloud-storage
gunicorn
